moves api into a subfolder
